pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dbporter"
version = "0.1.0"
description = "A powerful database migration tool with DAG support, automatic rollback, and schema inspection"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
maintainers = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = [
    "database", "migration", "schema", "sqlalchemy", "yaml", "dag", 
    "rollback", "postgresql", "mysql", "sqlite", "oracle", "sqlserver",
    "version-control", "database-versioning", "schema-management"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Database",
    "Topic :: System :: Archiving :: Backup",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Environment :: Console",
    "Framework :: SQLAlchemy",
]
requires-python = ">=3.8"
dependencies = [
    "typer>=0.9.0",
    "sqlalchemy>=2.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "tabulate>=0.9.0",
    "typing-extensions>=4.0.0",
]

[project.optional-dependencies]
postgresql = ["psycopg2-binary>=2.9.0"]
mysql = ["PyMySQL>=1.0.0"]
sqlserver = ["pyodbc>=4.0.0"]
oracle = ["cx-Oracle>=8.0.0"]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "isort>=5.12.0",
    "sphinx>=6.0.0",
    "sphinx-rtd-theme>=1.2.0",
    "pre-commit>=3.0.0",
    "ipython>=8.0.0",
    "jupyter>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/dbPorter"
Documentation = "https://github.com/yourusername/dbPorter#readme"
Repository = "https://github.com/yourusername/dbPorter.git"
"Bug Tracker" = "https://github.com/yourusername/dbPorter/issues"

[project.scripts]
dbporter = "dbPorter.main:main"
db-porter = "dbPorter.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["dbPorter*"]

[tool.setuptools.package-data]
dbPorter = ["examples/*.py", "*.md", "*.txt"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["dbPorter"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
    "--cov=dbPorter",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
